+++ lib/src/blas/cgemm/cgemm_array_kernels_batch.cpp
@@ -82,8 +82,8 @@ hcblasStatus cgemm_NoTransAB_batch_loopunroll(hc::accelerator_view accl_view,
     if (Row < N && Col < M) {
       CReal = C[elt][cOffset + (tidx.global[1] * ldc) + tidx.global[2]].x;
       CImg = C[elt][cOffset + (tidx.global[1] * ldc) + tidx.global[2]].y;
-      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
       tempReal = ((CReal * beta.x) - (CImg * beta.y));
       tempImg = ((CReal * beta.y) + (CImg * beta.x));
       C[elt][cOffset + (tidx.global[1] * ldc) + tidx.global[2]].x = tempReal + ((CValue * alpha.x) - (CValue1 * alpha.y));
@@ -104,14 +104,14 @@ hcblasStatus cgemm_NoTransAB_batch_MICRO_TS16XMTS2(hc::accelerator_view accl_vie
 					           float_2 alpha, float_2 beta, int batchSize) {
 #define TILESIZE 16
 #define MICROTILESIZE 1
-  int M_ = hc::fast_math::fmax(1, ((M + 1) / MICROTILESIZE));
-  int N_ = hc::fast_math::fmax(1, ((N + 1) / MICROTILESIZE));
+  int M_ = hc::fast_math::fmaxf(1, ((M + 1) / MICROTILESIZE));
+  int N_ = hc::fast_math::fmaxf(1, ((N + 1) / MICROTILESIZE));
   hc::extent<3> grdExt(batchSize, (N_ + (TILESIZE - 1)) & ~(TILESIZE - 1), (M_ + (TILESIZE - 1)) & ~(TILESIZE - 1));
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftTS = hc::fast_math::log2(TILESIZE);
-    int shiftMTP = hc::fast_math::log2(MICROTILEPROD);
+    int shiftTS = hc::fast_math::log2f(TILESIZE);
+    int shiftMTP = hc::fast_math::log2f(MICROTILEPROD);
     float rCreal[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rCimg[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rAreal[1][MICROTILESIZE] = {{(float)0}};
@@ -181,8 +181,8 @@ hcblasStatus cgemm_NoTransAB_batch_MICRO_TS16XMTS2(hc::accelerator_view accl_vie
         if(xIndex + (col << shiftTS) < M && (yIndex / ldc) + (row << shiftTS) < N) {
           CReal = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x;
           CImg = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row * TILESIZE) * ldc].y;
-          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
           tempReal = ((CReal * beta.x) - (CImg * beta.y));
           tempImg  = ((CReal * beta.y) + (CImg * beta.x));
           C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x = tempReal + ((rCreal[col][row] * alpha.x) - (rCimg[col][row] * alpha.y));
@@ -210,7 +210,7 @@ hcblasStatus cgemm_NoTransAB_batch_STEP_TS8XSS8(hc::accelerator_view accl_view,
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftFactor = hc::fast_math::log2(STEPSIZE);
+    int shiftFactor = hc::fast_math::log2f(STEPSIZE);
     float rCreal[1][1];
     float rAreal[1][STEPSIZE / TILESIZE];
     float rBreal[1][STEPSIZE / TILESIZE];
@@ -278,8 +278,8 @@ hcblasStatus cgemm_NoTransAB_batch_STEP_TS8XSS8(hc::accelerator_view accl_view,
     if(gidx * TILESIZE + idx < M && gidy * TILESIZE + idy < N) {
       CReal = C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy) * ldc].x;
       CImg = C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy) * ldc].y;
-      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
       tempReal = ((CReal * beta.x) - (CImg * beta.y));
       tempImg  = ((CReal * beta.y) + (CImg * beta.x));
       C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy)*ldc].x = tempReal + ((rCreal[0][0] * alpha.x) - (rCimg[0][0] * alpha.y));
@@ -301,14 +301,14 @@ hcblasStatus cgemm_NoTransAB_batch_MICRO_TS8XMTS2(hc::accelerator_view accl_view
 					          float_2 alpha, float_2 beta, int batchSize) {
 #define TILESIZE 8
 #define MICROTILESIZE 1
-  int M_ = hc::fast_math::fmax(1, ((M + 1) / MICROTILESIZE));
-  int N_ = hc::fast_math::fmax(1, ((N + 1)/ MICROTILESIZE));
+  int M_ = hc::fast_math::fmaxf(1, ((M + 1) / MICROTILESIZE));
+  int N_ = hc::fast_math::fmaxf(1, ((N + 1)/ MICROTILESIZE));
   hc::extent<3> grdExt(batchSize, (N_ + (TILESIZE - 1)) & ~(TILESIZE - 1), (M_ + (TILESIZE - 1)) & ~(TILESIZE - 1));
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftTS = hc::fast_math::log2(TILESIZE);
-    int shiftMTP = hc::fast_math::log2(MICROTILEPROD);
+    int shiftTS = hc::fast_math::log2f(TILESIZE);
+    int shiftMTP = hc::fast_math::log2f(MICROTILEPROD);
     float rCreal[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rCimg[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rAreal[1][MICROTILESIZE] = {{(float)0}};
@@ -378,8 +378,8 @@ hcblasStatus cgemm_NoTransAB_batch_MICRO_TS8XMTS2(hc::accelerator_view accl_view
         if(xIndex + (col << shiftTS) < M && (yIndex / ldc) + (row << shiftTS) < N) {
           CReal = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x;
           CImg = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row * TILESIZE) * ldc].y;
-          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
           tempReal = ((CReal * beta.x) - (CImg * beta.y));
           tempImg  = ((CReal * beta.y) + (CImg * beta.x));
           C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x = tempReal + ((rCreal[col][row] * alpha.x) - (rCimg[col][row] * alpha.y));
@@ -401,14 +401,14 @@ hcblasStatus cgemm_NoTransA_batch_MICRO_TS16XMTS2(hc::accelerator_view accl_view
 					          float_2 alpha, float_2 beta, int batchSize) {
 #define TILESIZE 16
 #define MICROTILESIZE 1
-  int M_ = hc::fast_math::fmax(1, ((M + 1) / MICROTILESIZE));
-  int N_ = hc::fast_math::fmax(1, ((N + 1) / MICROTILESIZE));
+  int M_ = hc::fast_math::fmaxf(1, ((M + 1) / MICROTILESIZE));
+  int N_ = hc::fast_math::fmaxf(1, ((N + 1) / MICROTILESIZE));
   hc::extent<3> grdExt(batchSize, (N_ + (TILESIZE - 1)) & ~(TILESIZE - 1), (M_ + (TILESIZE - 1)) & ~(TILESIZE - 1));
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftTS = hc::fast_math::log2(TILESIZE);
-    int shiftMTP = hc::fast_math::log2(MICROTILEPROD);
+    int shiftTS = hc::fast_math::log2f(TILESIZE);
+    int shiftMTP = hc::fast_math::log2f(MICROTILEPROD);
     float rCreal[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rCimg[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rAreal[1][MICROTILESIZE] = {{(float)0}};
@@ -478,8 +478,8 @@ hcblasStatus cgemm_NoTransA_batch_MICRO_TS16XMTS2(hc::accelerator_view accl_view
         if(xIndex + (col << shiftTS) < M && (yIndex / ldc) + (row << shiftTS) < N) {
           CReal = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x;
           CImg = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row * TILESIZE) * ldc].y;
-          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
           tempReal = ((CReal * beta.x) - (CImg * beta.y));
           tempImg  = ((CReal * beta.y) + (CImg * beta.x));
           C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x = tempReal + ((rCreal[col][row] * alpha.x) - (rCimg[col][row] * alpha.y));
@@ -505,7 +505,7 @@ hcblasStatus cgemm_NoTransB_batch_STEP_TS8XSS8(hc::accelerator_view accl_view,
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftFactor = hc::fast_math::log2(TILESIZE);
+    int shiftFactor = hc::fast_math::log2f(TILESIZE);
     float rCreal[1][1];
     float rAreal[1][1];
     float rBreal[1][1];
@@ -570,8 +570,8 @@ hcblasStatus cgemm_NoTransB_batch_STEP_TS8XSS8(hc::accelerator_view accl_view,
     if(gidx * TILESIZE + idx < M && gidy * TILESIZE + idy < N) {
       CReal = C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy) * ldc].x;
       CImg = C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy) * ldc].y;
-      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
       tempReal = ((CReal * beta.x) - (CImg * beta.y));
       tempImg  = ((CReal * beta.y) + (CImg * beta.x));
       C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy)*ldc].x = tempReal + ((rCreal[0][0] * alpha.x) - (rCimg[0][0] * alpha.y));
@@ -591,14 +591,14 @@ hcblasStatus cgemm_NoTransB_batch_MICRO_TS16XMTS2(hc::accelerator_view accl_view
 					          float_2 alpha, float_2 beta, int batchSize) {
 #define TILESIZE 16
 #define MICROTILESIZE 1
-  int M_ = hc::fast_math::fmax(1, ((M + 1) / MICROTILESIZE));
-  int N_ = hc::fast_math::fmax(1, ((N + 1) / MICROTILESIZE));
+  int M_ = hc::fast_math::fmaxf(1, ((M + 1) / MICROTILESIZE));
+  int N_ = hc::fast_math::fmaxf(1, ((N + 1) / MICROTILESIZE));
   hc::extent<3> grdExt(batchSize, (N_ + (TILESIZE - 1)) & ~(TILESIZE - 1), (M_ + (TILESIZE - 1)) & ~(TILESIZE - 1));
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftTS = hc::fast_math::log2(TILESIZE);
-    int shiftMTP = hc::fast_math::log2(MICROTILEPROD);
+    int shiftTS = hc::fast_math::log2f(TILESIZE);
+    int shiftMTP = hc::fast_math::log2f(MICROTILEPROD);
     float rCreal[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rCimg[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rAreal[1][MICROTILESIZE] = {{(float)0}};
@@ -669,8 +669,8 @@ hcblasStatus cgemm_NoTransB_batch_MICRO_TS16XMTS2(hc::accelerator_view accl_view
         if(xIndex + (col << shiftTS) < M && (yIndex / ldc) + (row << shiftTS) < N) {
           CReal = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x;
           CImg = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row * TILESIZE) * ldc].y;
-          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
           tempReal = ((CReal * beta.x) - (CImg * beta.y));
           tempImg  = ((CReal * beta.y) + (CImg * beta.x));
           C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x = tempReal + ((rCreal[col][row] * alpha.x) - (rCimg[col][row] * alpha.y));
@@ -765,8 +765,8 @@ hcblasStatus cgemm_NoTransB_batch_loopunroll(hc::accelerator_view accl_view,
     if (Row < N && Col < M) {
       CReal = C[elt][cOffset + (tidx.global[1] * ldc) + tidx.global[2]].x;
       CImg = C[elt][cOffset + (tidx.global[1] * ldc) + tidx.global[2]].y;
-      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
       tempReal = ((CReal * beta.x) - (CImg * beta.y));
       tempImg = ((CReal * beta.y) + (CImg * beta.x));
       C[elt][cOffset + (tidx.global[1] * ldc) + tidx.global[2]].x = tempReal + ((CValue * alpha.x) - (CValue1 * alpha.y));
@@ -790,7 +790,7 @@ hcblasStatus cgemm_TransAB_batch_STEP_TS8XSS8(hc::accelerator_view accl_view,
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftFactor = hc::fast_math::log2(TILESIZE);
+    int shiftFactor = hc::fast_math::log2f(TILESIZE);
     float rCreal[1][1];
     float rAreal[1][1];
     float rBreal[1][1];
@@ -852,8 +852,8 @@ hcblasStatus cgemm_TransAB_batch_STEP_TS8XSS8(hc::accelerator_view accl_view,
     if(gidx * TILESIZE + idx < M && gidy * TILESIZE + idy < N) {
       CReal = C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy) * ldc].x;
       CImg = C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy) * ldc].y;
-      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
       tempReal = ((CReal * beta.x) - (CImg * beta.y));
       tempImg  = ((CReal * beta.y) + (CImg * beta.x));
       C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy)*ldc].x = tempReal + ((rCreal[0][0] * alpha.x) - (rCimg[0][0] * alpha.y));
@@ -877,7 +877,7 @@ hcblasStatus cgemm_TransAB_batch_STEP_TS16XSS16(hc::accelerator_view accl_view,
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftFactor = hc::fast_math::log2(TILESIZE);
+    int shiftFactor = hc::fast_math::log2f(TILESIZE);
     float rCreal[1][1];
     float rAreal[1][1];
     float rBreal[1][1];
@@ -939,8 +939,8 @@ hcblasStatus cgemm_TransAB_batch_STEP_TS16XSS16(hc::accelerator_view accl_view,
     if(gidx * TILESIZE + idx < M && gidy * TILESIZE + idy < N) {
       CReal = C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy) * ldc].x;
       CImg = C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy) * ldc].y;
-      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+      CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+      CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
       tempReal = ((CReal * beta.x) - (CImg * beta.y));
       tempImg  = ((CReal * beta.y) + (CImg * beta.x));
       C[elt][cOffset + gidx * TILESIZE + idx + (gidy * TILESIZE + idy)*ldc].x = tempReal + ((rCreal[0][0] * alpha.x) - (rCimg[0][0] * alpha.y));
@@ -960,14 +960,14 @@ hcblasStatus cgemm_TransAB_batch_MICRO_TS16XMTS2(hc::accelerator_view accl_view,
 					         float_2 alpha, float_2 beta, int batchSize) {
 #define TILESIZE 16
 #define MICROTILESIZE 1
-  int M_ = hc::fast_math::fmax(1, ((M + 1) / MICROTILESIZE));
-  int N_ = hc::fast_math::fmax(1, ((N + 1) / MICROTILESIZE));
+  int M_ = hc::fast_math::fmaxf(1, ((M + 1) / MICROTILESIZE));
+  int N_ = hc::fast_math::fmaxf(1, ((N + 1) / MICROTILESIZE));
   hc::extent<3> grdExt(batchSize, (N_ + (TILESIZE - 1)) & ~(TILESIZE - 1), (M_ + (TILESIZE - 1)) & ~(TILESIZE - 1));
   hc::tiled_extent<3> t_ext = grdExt.tile(1, TILESIZE, TILESIZE);
   hc::parallel_for_each(accl_view, t_ext, [ = ] (hc::tiled_index<3> tidx) [[hc]] {
     int elt = tidx.tile[0];
-    int shiftTS = hc::fast_math::log2(TILESIZE);
-    int shiftMTP = hc::fast_math::log2(MICROTILEPROD);
+    int shiftTS = hc::fast_math::log2f(TILESIZE);
+    int shiftMTP = hc::fast_math::log2f(MICROTILEPROD);
     float rCreal[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rCimg[MICROTILESIZE][MICROTILESIZE] = {{(float)0}};
     float rAreal[1][MICROTILESIZE] = {{(float)0}};
@@ -1037,8 +1037,8 @@ hcblasStatus cgemm_TransAB_batch_MICRO_TS16XMTS2(hc::accelerator_view accl_view,
         if(xIndex + (col << shiftTS) < M && (yIndex / ldc) + (row << shiftTS) < N) {
           CReal = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x;
           CImg = C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row * TILESIZE) * ldc].y;
-          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(CReal)) ? 0 : CReal;
-          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(CImg)) ? 0 : CImg;
+          CReal = (hc::fast_math::isnan(CReal) || hc::fast_math::isinf(static_cast<float>(CReal))) ? 0 : CReal;
+          CImg = (hc::fast_math::isnan(CImg) || hc::fast_math::isinf(static_cast<float>(CImg))) ? 0 : CImg;
           tempReal = ((CReal * beta.x) - (CImg * beta.y));
           tempImg  = ((CReal * beta.y) + (CImg * beta.x));
           C[elt][cOffset + (xIndex + (col << shiftTS)) + yIndex + (row << shiftTS) * ldc].x = tempReal + ((rCreal[col][row] * alpha.x) - (rCimg[col][row] * alpha.y));
